Detect the swung object automatically, and confirm it before rendering
The first click is now optional. `lightsaber-fx run` looks for the object
itself, shows you what it found with the mask overlaid, and takes ENTER as
acceptance or a click as an override. `--no-auto` skips the search.

Motion decides *where* to look, SAM2 decides *what is there*, and
`fit_blade` decides whether that is blade-like enough to propose. The
object you want is the thing being *swung*, which is a far more specific
target than "a bat-shaped thing" -- and it is what rules out the
distractors that defeat a shape-only search. A fence rail, the horizon and
a chalk line are all longer and straighter than any bat, and all perfectly
still, so they score zero no matter how bat-shaped they look. Camera pans
are handled by subtracting the median flow vector; zoom and roll are not,
and that is written down as a limit rather than left to be discovered.

The first version of this scored elongation on the optical-flow region and
gated on it, which measures the wrong thing. Flow magnitude on a rotating
object scales with radius, so a bat pivoting about the hands lights up only
near its tip; the resulting blob has an elongation around 2, and any gate
set sensibly for a blade rejects it. A rotation-dominated swing would
silently never have been detected. The synthetic rotating-bar fixture in
test_detect.py exists to pin that case, and it is why the shape judgement
moved onto the SAM2 mask -- which is the actual object, and is also exactly
what the renderer later has to draw. On the 10 s clip that raised the
proposed mask from elongation 5.0 (motion blob, barrel only) to 10.1 (the
whole bat).

A percentile is a relative threshold, so on a clip where nothing moves it
lands in the codec noise and cheerfully reports the "fastest" 3% of a still
frame: measured, a 2318-pixel component at a mean speed of 0.00. An
absolute floor is what lets "nothing is moving" be answered as such.

Tracking now propagates in both directions from the prompt frame, because
the frame where a swung object is easiest to find is rarely frame 0 -- it is
135 of 300 on the 10 s clip. Verified end to end on real footage: detection
took 2.0 s, and all 60 masks came back non-empty including all 21 frames
*before* the prompt frame, with comparable median areas either side
(2544 before, 2352 from the prompt on).

Detection never emits an exclude point. An exclude that lands on the object
is how a track ends up empty, and every proposed point is chosen as a real
foreground pixel of the mask rather than constructed geometrically -- a
centroid of a bent object can fall off it.

Measured on the test clips: 2.0 s and 4.1 s, CPU-only, so it does not
contend with tracking for the GPU. No new dependencies.

220 tests pass. Repo-wide ruff findings also drop from 21 to 14 (redundant
int(round(...)) casts, an unused typing.Optional, import ordering, and an
implicit naive-local datetime in `jobs` made explicit).

+++ src/lightsaber_fx/pipeline/audio.py
@@ -111,7 +111,7 @@ def _voice_params(voice):
 # ---------------------------------------------------------------------------
 
 def _n_samples(duration, sr=SR):
-    return max(0, int(round(sr * duration)))
+    return max(0, round(sr * duration))
 
 
 def _osc(freq_t, sr=SR):
@@ -255,7 +255,7 @@ def _interp_and_smooth(frame_values, fps, n_samples, sr=SR, smooth_ms=60.0):
     # returns a length-max(M, N) array when the kernel is longer than the
     # signal, which would silently break every caller's length assumption
     # on short clips.
-    win = min(max(1, int(round(smooth_ms / 1000.0 * sr))), n_samples)
+    win = min(max(1, round(smooth_ms / 1000.0 * sr)), n_samples)
     if win > 1 and n_samples > 1:
         kernel = np.ones(win, dtype=np.float64) / win
         interped = np.convolve(interped, kernel, mode="same")
@@ -412,7 +412,7 @@ def synth_tv_buzz(duration, voice="neutral", seed=1, sr=SR):
     if peak > 0:
         buzz = buzz / peak
 
-    n_crackles = max(1, int(round(duration * 3)))
+    n_crackles = max(1, round(duration * 3))
     crackle_len = max(1, int(0.003 * sr))
     crackle_track = np.zeros(n, dtype=np.float64)
     for pos in rng.integers(0, n, size=n_crackles):
